improve(scenery): move all FP three helpers class in a new directory with more explicit names
